working on fixing all the errors from switching to postgres, especially replacing text with UUID, cleaning up some logging along the way aswell. transactions seems to be cleaning up, will workon setting up postgresql container test environment and update my table testing accordingly

+++ cache/redis.go
@@ -33,14 +33,14 @@ func WarmCategoriesCache(cfg *config.ApiConfig) error {
 		return err
 	}
 
-	if err = storeCategoriesAsHash(ctx, RedisClient, "primary_categories", primaryCategories, func(p models.PrimaryCategory) int64 {
+	if err = storeCategoriesAsHash(ctx, RedisClient, "primary_categories", primaryCategories, func(p models.PrimaryCategory) int32 {
 		return p.ID
 	}); err != nil {
 		log.Printf("error hashing primary_categories into the cache: %v", err)
 		return err
 	}
 
-	if err = storeCategoriesAsHash(ctx, RedisClient, "detailed_categories", detailedCategories, func(d models.DetailedCategory) int64 {
+	if err = storeCategoriesAsHash(ctx, RedisClient, "detailed_categories", detailedCategories, func(d models.DetailedCategory) int32 {
 		return d.ID
 	}); err != nil {
 		log.Printf("error hashing detailed_categories into the cache: %v", err)
@@ -51,14 +51,14 @@ func WarmCategoriesCache(cfg *config.ApiConfig) error {
 	return nil
 }
 
-func storeCategoriesAsHash[T any](ctx context.Context, client *redis.Client, keyName string, categories []T, idExtractor func(T) int64) error {
+func storeCategoriesAsHash[T any](ctx context.Context, client *redis.Client, keyName string, categories []T, idExtractor func(T) int32) error {
 	redisHash := make(map[string]interface{}, len(categories))
 	for _, cat := range categories {
 		catJSON, err := json.Marshal(cat)
 		if err != nil {
 			return err
 		}
-		fieldName := strconv.FormatInt(idExtractor(cat), 10)
+		fieldName := strconv.FormatInt(int64(idExtractor(cat)), 10)
 		redisHash[fieldName] = string(catJSON)
 	}
 	return client.HSet(ctx, keyName, redisHash).Err()
